+++ CHANGES.md
@@ -178,6 +178,10 @@ let mut dataset = driver
 
   - <https://github.com/georust/gdal/pull/200>
 
+- Wrappers for `CPLSetThreadLocalConfigOption` and `CPLGetThreadLocalConfigOption`
+
+  - <https://github.com/georust/gdal/pull/201>
+
 ## 0.7.1
 
 - fix docs.rs build for gdal-sys